requirements.txt
Flask
Flask-Login
requests
python-dotenv
